backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
numpy>=1.26.0
opencv-python-headless==4.10.0.84
Pillow>=10.2.0
python-multipart==0.0.20
aiohttp>=3.9.0
modal>=0.67.0
ultralytics>=8.3.0
